requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.37.0
accelerate>=0.24.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=9.5.0
scikit-learn>=1.3.0
gradio>=4.0.0
# sqlite3 is built-in to Python
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0